database stuff

+++ internal/types/types.go
@@ -1,6 +1,6 @@
 package types
 
 type Stats struct {
-	RX_bytes int
-	TX_bytes int
+	RX_bytes int64
+	TX_bytes int64
 }
